Changes: - migrations:   - modify contraints actions (on update/on delete).   - add table reports. - front-app:   - add post report system.   - fix few bugs.   - improve UI/UX. - admin:   - add filter on person list.   - add list of reported post on dashboard page.

+++ resources/app/js/components/question/List.vue
@@ -23,14 +23,26 @@
                                 </p>
                             </v-flex>
                             <v-spacer></v-spacer>
-                            <v-menu v-show="item.user_id == user.id">
+
+                            <v-menu>
                                 <v-btn slot="activator" icon>
                                     <v-icon>more_vert</v-icon>
                                 </v-btn>
                                 <v-list>
-                                    <v-list-tile @click="deletePost(item.id)">
-                                        <v-list-tile-title>
+                                    <v-list-tile 
+                                        @click="deletePost(item.id)"
+                                        v-show="item.user_id == user.id">
+                                        <v-list-tile-title class="body-1">
+                                            <v-icon>close</v-icon>
                                             {{ $t('action.delete') }}
+                                        </v-list-tile-title>
+                                    </v-list-tile>
+                                    <v-list-tile 
+                                        @click="showReportPost(item)"
+                                        v-show="item.user_id != user.id">
+                                        <v-list-tile-title class="body-1">
+                                            <v-icon>report</v-icon>
+                                            {{ $t('action.report') }}
                                         </v-list-tile-title>
                                     </v-list-tile>
                                 </v-list>
@@ -47,6 +59,43 @@
                         </v-card-title>
                     </v-card>
                 </v-flex>
+
+                <el-dialog
+                    title=""
+                    :visible.sync="report.show"
+                    width="40%"
+                    center>
+                    <span>
+                        <!-- show this if logged user not yet report this post -->
+                        <v-flex v-show="!report.post.reported" lg12 v-for="(reason, index) in reasons">
+                            <el-radio 
+                                v-model="report.i_reason"
+                                :label="index">
+                                {{ reason }}
+                            </el-radio>
+                            <el-input 
+                                v-if="index + 1 == reasons.length" 
+                                v-show="report.i_reason == reasons.length - 1"
+                                v-model="report.reason">
+                            </el-input>
+                        </v-flex>
+                        <!-- show this if logged user not yet report this post -->
+
+                        <!-- show this if logged user already report this post -->
+                        <v-flex v-show="report.post.reported" lg12>
+                            Reported for <i>"{{ report.post.reported }}"</i>
+                        </v-flex>
+                        <!-- show this if logged user already report this post -->
+                    </span>
+
+                    <span slot="footer" class="dialog-footer" v-show="!report.post.reported">
+                        <el-button @click="cancelReportPost()">{{ $t('action.cancel') }}</el-button>
+                        <el-button type="primary" @click="reportPost()">{{ $t('action.report') }}</el-button>
+                    </span>
+                    <span slot="footer" class="dialog-footer" v-show="report.post.reported">
+                        <el-button @click="cancelReportPost()">Ok</el-button>
+                    </span>
+                </el-dialog>
             </v-layout>
         </template>
     </div>
@@ -56,22 +105,89 @@
     export default {
         name: "Question",
         props: { data: Array },
+        data() {
+            return {
+                report: {
+                    post: {},
+                    reason: null,
+                    show: false,
+                    i_reason: null,
+                },
+                reasons: [
+                    "Uncivil, unneighborly or offensive",
+                    "Not relevant or annoying",
+                    "Safety issue or illegal", 
+                    "Commercial or spam", 
+                    "Others",
+                ],
+            };
+        },
         computed: {
             user() {
                 return this.$store.state.auth.user;
             }
         },
         methods: {
-            async deletePost(post_id) {
+            showReportPost(post) {
+                this.report.show = true;
+                this.report.i_reason = null;
+                this.report.reason = null;
+                this.report.post = post;
+            },
+            cancelReportPost(post_id) {
+                this.report.show = false;
+
+                // just delay the states change
+                setTimeout(() => {
+                    this.report.post = {};
+                    this.report.i_reason = null;
+                    this.report.reason = null;
+                }, 500);
+            },
+            async reportPost() {
+                let params = {
+                    post_id: this.report.post.id,
+                    reason: this.report.reason,
+                };
+
+                if (this.report.i_reason < this.reasons.length -1) 
+                    params.reason = this.reasons[this.report.i_reason];
+
                 // delete post
-                let response = await this.$http.delete("/question/" + post_id)
+                let response = await this.$http.post("/question/report", params)
                     .then(() => {
                         this.$emit("refresh-list");
-                        this.$message.error(this.$t("question.messages.delete_success"));
+
+                        this.report.post.reported = this.report.reason;
+                        this.$message.success(this.$t("question.messages.report_success"));
+
+                        // reset all
+                        this.cancelReportPost()
                     })
                     .catch((err) => {
                         this.$message.error(this.$t("question.messages.action_failed"));
+                        // reset all
+                        this.cancelReportPost()
                     });
+
+            },
+
+            async deletePost(post_id) {
+                this.$confirm(this.$t('question.labels.delete_confirmation'), 'Warning', {
+                    confirmButtonText: this.$t('action.sure'),
+                    cancelButtonText: this.$t('action.cancel'),
+                    type: 'warning'
+                }).then(async () => {
+                    // delete post
+                    let response = await this.$http.delete("/question/" + post_id)
+                        .then(() => {
+                            this.$emit("refresh-list");
+                            this.$message.success(this.$t("question.messages.delete_success"));
+                        })
+                        .catch((err) => {
+                            this.$message.error(this.$t("question.messages.action_failed"));
+                        });
+                });
             },
         },
     }
